refactor(agents-server-ui): dedupe formatSender, move comment helpers to lib
formatSender/userDisplayName live once in lib/principals.ts (CommentBubble
had re-implemented UserMessage's copy with drifting ellipsis). The pure
comment-target codecs, buildCommentsTimeline, and TimelineRowAdjacency move
from ChatView.tsx to lib/comments.ts with their tests.
