+++ packages/agents-server-ui/src/components/views/ChatView.tsx
@@ -3,14 +3,19 @@ import { useNavigate } from '@tanstack/react-router'
 import { eq, useLiveQuery } from '@tanstack/react-db'
 import { useEntityTimeline } from '../../hooks/useEntityTimeline'
 import { useForkFromHere } from '../../hooks/useForkFromHere'
-import { EntityTimeline, type TimelineRowAdjacency } from '../EntityTimeline'
+import { EntityTimeline } from '../EntityTimeline'
 import { GoalBanner } from '../GoalBanner'
 import { MessageInput } from '../MessageInput'
 import { EntityContextDrawer } from '../EntityContextDrawer'
 import { useElectricAgents } from '../../lib/ElectricAgentsProvider'
 import { useWorkspace } from '../../hooks/useWorkspace'
-import { isAttachmentManifest } from '../../lib/attachments'
 import { schemaModelSupportsImageInput } from '../../lib/modelCapabilities'
+import {
+  buildCommentsTimeline,
+  COMMENT_FOCUS_PARAM,
+  commentFocusViewParams,
+  decodeCommentTargetParam,
+} from '../../lib/comments'
 import type { SelectedCommentTarget, TimelineRow } from '../../lib/comments'
 import {
   useEntityPermission,
@@ -27,94 +32,6 @@ const CHAT_VIEW_PERMISSIONS: ReadonlyArray<EntityPermission> = [
   `signal`,
   `fork`,
 ]
-const COMMENT_FOCUS_PARAM = `focus`
-const COMMENT_TARGET_COLLECTIONS = new Set<string>([
-  `inbox`,
-  `run`,
-  `text`,
-  `tool_call`,
-  `wake`,
-  `signal`,
-  `manifest`,
-])
-
-export function encodeCommentTargetParam(target: CommentTarget): string {
-  return encodeURIComponent(JSON.stringify(target))
-}
-
-export function decodeCommentTargetParam(
-  value: string | undefined
-): CommentTarget | null {
-  if (!value) return null
-  try {
-    const decoded = JSON.parse(decodeURIComponent(value)) as unknown
-    if (!isCommentTarget(decoded)) return null
-    return decoded
-  } catch {
-    return null
-  }
-}
-
-export function commentFocusViewParams(
-  target: CommentTarget
-): Record<string, string> {
-  return { [COMMENT_FOCUS_PARAM]: encodeCommentTargetParam(target) }
-}
-
-function isCommentTarget(value: unknown): value is CommentTarget {
-  if (!value || typeof value !== `object`) return false
-  const target = value as Partial<CommentTarget>
-  if (target.kind === `comment`) {
-    return typeof target.key === `string`
-  }
-  if (target.kind !== `timeline`) return false
-  const timelineTarget = target as Partial<
-    Extract<CommentTarget, { kind: `timeline` }>
-  >
-  return (
-    typeof timelineTarget.key === `string` &&
-    typeof timelineTarget.collection === `string` &&
-    COMMENT_TARGET_COLLECTIONS.has(timelineTarget.collection) &&
-    (timelineTarget.run_id === undefined ||
-      typeof timelineTarget.run_id === `string`)
-  )
-}
-
-export function buildCommentsTimeline(timelineRows: Array<TimelineRow>): {
-  rows: Array<TimelineRow>
-  adjacency: Array<TimelineRowAdjacency>
-} {
-  const rows: Array<TimelineRow> = []
-  const adjacency: Array<TimelineRowAdjacency> = []
-  let previousRenderableRow: TimelineRow | undefined
-  let pendingCommentAdjacencyIndex: number | null = null
-
-  for (const row of timelineRows) {
-    if (isAttachmentManifest(row.manifest)) continue
-
-    if (pendingCommentAdjacencyIndex !== null) {
-      const pendingAdjacency = adjacency[pendingCommentAdjacencyIndex]!
-      adjacency[pendingCommentAdjacencyIndex] = {
-        ...pendingAdjacency,
-        nextRow: row,
-      }
-      pendingCommentAdjacencyIndex = null
-    }
-
-    if (row.comment) {
-      rows.push(row)
-      adjacency.push({
-        previousRow: previousRenderableRow,
-      })
-      pendingCommentAdjacencyIndex = adjacency.length - 1
-    }
-
-    previousRenderableRow = row
-  }
-
-  return { rows, adjacency }
-}
-
 /**
  * The default view: chat / timeline + message composer.
  *
